@@ -13,21 +13,23 @@
   },
   "devDependencies": {
     "@parcel/transformer-less": "^2.0.1",
-    "@playwright/test": "^1.20.1",
+    "@playwright/test": "1.16.1",
     "@swc/core": "^1.2.138",
     "@swc/jest": "^0.2.17",
-    "@testing-library/react": "^12.1.3",
-    "@testing-library/user-event": "^14.0.0-beta",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.1.1",
+    "@testing-library/user-event": "^14.1.1",
     "@types/jest": "^27.0.2",
+    "@types/jsdom": "^16.2.14",
     "@types/lodash": "^4.14.177",
     "@types/ramda": "^0.27.57",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
-    "@types/react-test-renderer": "^17.0.1",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "@types/react-test-renderer": "^18.0.0",
     "jest": "^27.5.0",
     "parcel": "^2.0.1",
-    "react-test-renderer": "^17.0.2",
-    "ts-jest": "^27.0.7",
+    "react-test-renderer": "^18.0.0",
+    "ts-jest": "^27.1.4",
     "typescript": "^4.4.4"
   },
   "dependencies": {
@@ -39,8 +41,8 @@
     "lodash": "^4.17.21",
     "ramda": "^0.27.1",
     "rc-tree": "^5.5.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-error-boundary": "^3.1.4",
     "react-split-pane": "^0.1.92",
     "recast": "^0.20.5",
