@@ -1,19 +1,22 @@
-import {updateState} from './appData'
+const appFunctions = (state: {_updateApp: (changes: object) => void }) => {
+    const {_updateApp} = state
+    return ({
+        ShowPage(pageName: string) {
+            _updateApp( {_data: {currentPage: pageName}})
+        },
 
-export default {
-    ShowPage(pageName: string) {
-        updateState('app._data', {currentPage: pageName})
-    },
+        Reset(component: {_update: (changes: {value: any}) => void}) {
+            component._update({value: undefined})
+        },
 
-    Reset(component: {_update: (changes: {value: any}) => void}) {
-        component._update({value: undefined})
-    },
+        Set(component: {_update: (value: any, replace?: boolean) => void}, value: any) {
+            component._update({value}, true)
+        },
 
-    Set(component: {_update: (value: any, replace?: boolean) => void}, value: any) {
-        component._update({value}, true)
-    },
-
-    Update(component: {_update: (changes: object) => void}, changes: object) {
-        component._update({value: changes})
-    },
+        Update(component: {_update: (changes: object) => void}, changes: object) {
+            component._update({value: changes})
+        },
+    })
 }
+
+export default appFunctions
