@@ -4,18 +4,18 @@
 
 import React from 'react'
 import Editor from '../../src/editor/Editor'
-import {fireEvent, render, screen, within} from '@testing-library/react'
+import {act, fireEvent, render, screen, within} from '@testing-library/react/pure'
 import {ElementType} from '../../src/model/Types'
 import {startCase} from 'lodash'
-import {actWait, clickExpandControlFn, suppressRcTreeJSDomError, treeItemLabels} from '../testutil/testHelpers'
+import {ex, stopSuppressingRcTreeJSDomError, suppressRcTreeJSDomError, treeItemLabels} from '../testutil/testHelpers'
 import Page from '../../src/model/Page'
 import TextInput from '../../src/model/TextInput'
 import App from '../../src/model/App'
-import {ex} from '../../src/util/helpers'
 import {projectFixture1, projectFixture2} from '../testutil/projectFixtures'
 import Project from '../../src/model/Project'
+import {treeExpandControlSelector} from './Selectors'
 
-let container: any = null
+let container: any = null, unmount: any
 
 const itemLabels = () => treeItemLabels(container)
 const clickExpandControl = (...indexes: number[]) => clickExpandControlFn(container)(...indexes)
@@ -26,17 +26,39 @@ const onPropertyChange = ()=> {}
 const onAction = jest.fn()
 const onInsert = ()=> '123'
 
+export const wait = (time: number): Promise<void> => new Promise(resolve => setInterval(resolve, time))
+export const actWait = async (testFn: () => void) => {
+    await act(async () => {
+        testFn()
+        await wait(20)
+    })
+}
+export const clickExpandControlFn = (container: any) => async (...indexes: number[]) => {
+    for (const index of indexes) await actWait(() => fireEvent.click(container.querySelectorAll(treeExpandControlSelector)[index]))
+}
+
 beforeAll(suppressRcTreeJSDomError)
+afterAll(stopSuppressingRcTreeJSDomError)
+
+afterEach( async () => await act(() => {
+    try{
+        unmount && unmount()
+    } catch(e: any) {
+        if (!e.message?.match(/Cannot read properties of null \(reading 'removeEventListener'\)/)) {
+            throw e
+        }
+    }
+}))
 
 test("renders tree with app elements",  async () => {
-    await actWait(() =>  ({container} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
+    await actWait(() =>  ({container, unmount} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
     await clickExpandControl(0, 1, 2)
     expect(container.querySelector('.MuiTypography-h6').textContent).toBe("Elemento Studio")
     expect(itemLabels()).toStrictEqual(['Project One', 'App One', 'Main Page', 'First Text', 'Second Text', 'Other Page'])
 })
 
 test('shows Text element selected in tree in property editor', async () => {
-    await actWait(() =>  ({container} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
+    await actWait(() =>  ({container, unmount} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
     await clickExpandControl(0, 1, 2)
 
     expect(itemLabels()).toStrictEqual(['Project One', 'App One', 'Main Page', 'First Text', 'Second Text', 'Other Page'])
@@ -48,7 +70,7 @@ test('shows Text element selected in tree in property editor', async () => {
 })
 
 test('property kind button state does not leak into other properties', async () => {
-    await actWait(() =>  ({container} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
+    await actWait(() =>  ({container, unmount} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
     await clickExpandControl(0, 1, 2)
 
     expect(itemLabels()).toStrictEqual(['Project One', 'App One', 'Main Page', 'First Text', 'Second Text', 'Other Page'])
@@ -65,7 +87,7 @@ test('property kind button state does not leak into other properties', async () 
 })
 
 test('shows TextInput element selected in tree in property editor', async () => {
-    await actWait(() =>  ({container} = render(<Editor project={projectFixture2()} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
+    await actWait(() =>  ({container, unmount} = render(<Editor project={projectFixture2()} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
     await clickExpandControl(0, 1, 3)
 
     expect(itemLabels()).toStrictEqual(['Project One', 'App One', 'Main Page', 'Other Page', 'Some Text', 'Another Text Input', 'Button 2'])
@@ -88,7 +110,7 @@ test('shows errors for properties', async () => {
             new TextInput('textInput_1', 'First Text Input', {initialValue: ex`"A text value" + `, maxLength: ex`BadName + 30`}),
         ]),
     ]) ])
-    await actWait(() =>  ({container} = render(<Editor project={projectWithErrors} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
+    await actWait(() =>  ({container, unmount} = render(<Editor project={projectWithErrors} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
     // await actWait(() =>  fireEvent.click(container.querySelectorAll(treeExpandControlSelector)[0]))
     await clickExpandControl(0, 1, 2)
 
@@ -110,7 +132,7 @@ test('shows allowed items in insert menu', async () => {
     const optionsShown = () => screen.queryByTestId('insertMenu') && within(screen.getByTestId('insertMenu')).queryAllByRole('menuitem').map( el => el.textContent)
     const warningMessage = () => screen.getByTestId('insertWarning')
 
-    await actWait(() =>  ({container} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
+    await actWait(() =>  ({container, unmount} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
     fireEvent.click(screen.getByText('Insert'))
     expect(optionsShown()).toBeNull()
     expect(warningMessage().textContent).toMatch(/Please select/)
@@ -135,7 +157,7 @@ test.each(['Text', 'TextInput', 'NumberInput','SelectInput', 'TrueFalseInput', '
         return notionalNewElementId
     }
 
-    await actWait(() =>  ({container} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
+    await actWait(() =>  ({container, unmount} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
     await clickExpandControl(0, 1, 2)
 
     fireEvent.click(screen.getByText('Second Text'))
@@ -155,7 +177,7 @@ test(`notifies insert of Page with item selected in tree and selects new item`, 
         return notionalNewElementId
     }
 
-    await actWait(() =>  ({container} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
+    await actWait(() =>  ({container, unmount} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
     await clickExpandControl(0, 1, 2)
 
     fireEvent.click(screen.getByText('Main Page'))
@@ -169,7 +191,7 @@ test(`notifies insert of Page with item selected in tree and selects new item`, 
 
 test('notifies open request', async () => {
     let opened: boolean = false
-    await actWait(() =>  ({container} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction} onOpen={() => opened = true}/>)))
+    await actWait(() =>  ({container, unmount} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction} onOpen={() => opened = true}/>)))
     fireEvent.click(screen.getByText('File'))
     fireEvent.click(screen.getByText('Open'))
     expect(opened).toBe(true)
@@ -177,14 +199,14 @@ test('notifies open request', async () => {
 
 test('notifies save request', async () => {
     let saved: boolean = false
-    await actWait(() =>  ({container} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction} onSave={() => saved = true}/>)))
+    await actWait(() =>  ({container, unmount} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction} onSave={() => saved = true}/>)))
     fireEvent.click(screen.getByText('File'))
     fireEvent.click(screen.getByText('Save'))
     expect(saved).toBe(true)
 })
 
 test(`notifies tree action with item selected in tree`, async () => {
-    await actWait(() =>  ({container} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
+    await actWait(() =>  ({container, unmount} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
     await clickExpandControl(0, 1, 2)
 
     await actWait(() => fireEvent.contextMenu(screen.getByText('Second Text')))
@@ -195,8 +217,8 @@ test(`notifies tree action with item selected in tree`, async () => {
 })
 
 test('has iframe for running app', async () => {
-    await actWait(() =>  ({container} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
+    await actWait(() =>  ({container, unmount} = render(<Editor project={project} onChange={onPropertyChange} onInsert={onInsert} onAction={onAction}/>)))
 
     const appFrame = container.querySelector('iframe[name="appFrame"]')
-    expect(appFrame.src).toMatch(/.*\/run\/index.html$/)
+    expect(appFrame.src).toMatch(/.*\/run\?editorPreview$/)
 })
