@@ -9,8 +9,7 @@ import Page from '../../src/model/Page'
 import Text from '../../src/model/Text'
 import TextInput from '../../src/model/TextInput'
 import Button from '../../src/model/Button'
-import {componentProps} from '../testutil/testHelpers'
-import {ex} from '../../src/util/helpers'
+import {componentProps, ex} from '../testutil/testHelpers'
 import NumberInput from '../../src/model/NumberInput'
 import TrueFalseInput from '../../src/model/TrueFalseInput'
 import SelectInput from '../../src/model/SelectInput'
