@@ -4,23 +4,28 @@ import {treeExpand, treeItem} from './playwrightHelpers'
 // Expects test server such as Parcel dev server running on port 1234
 const runtimeRootUrl = 'http://localhost:1234/studio'
 
-let openMainPage = async function (page: Page) {
+let openMainPage = async (page: Page) => {
     await page.click(treeExpand(0))
     await page.click(treeExpand(1))
     await page.click(treeExpand(2))
     expect(await page.textContent(treeItem(5))).toBe('Third Text')
 }
 
+const getAppFrame = async (page: Page) => {
+    await page.locator('iframe[name="appFrame"]').waitFor();
+    return page.frame('appFrame') as Frame
+}
+
 test('app shown in frame', async ({ page }) => {
     await page.goto(runtimeRootUrl)
-    const appFrame = page.frame('appFrame') as Frame
+    const appFrame = await getAppFrame(page)
     expect(await appFrame.textContent('p >> nth=2')).toBe('Start your program here...')
 
 })
 
 test('Selecting element in editor highlights in the running app', async ({ page }) => {
     await page.goto(runtimeRootUrl)
-    const appFrame = page.frame('appFrame') as Frame
+    const appFrame = await getAppFrame(page)
     const getOutlineStyle = async (selector: string) => page.evaluate((el: any) => getComputedStyle(el).outlineStyle, await appFrame.$(selector))
 
     expect(await appFrame.textContent('p >> nth=2')).toBe('Start your program here...')
@@ -39,7 +44,7 @@ test('Selecting element in editor highlights in the running app', async ({ page 
 
 test('Changes to app definition show immediately in the running app', async ({ page }) => {
     await page.goto(runtimeRootUrl)
-    const appFrame = page.frame('appFrame') as Frame
+    const appFrame = await getAppFrame(page)
     expect(await appFrame.textContent('p >> nth=2')).toBe('Start your program here...')
 
     await openMainPage(page)
@@ -55,7 +60,7 @@ test('Changes to app definition show immediately in the running app', async ({ p
 
 test('Formulas in app definition update the running app immediately', async ({ page })=> {
     await page.goto(runtimeRootUrl)
-    const appFrame = page.frame('appFrame') as Frame
+    const appFrame = await getAppFrame(page)
     expect(await appFrame.textContent('p >> nth=2')).toBe('Start your program here...')
 
     await openMainPage(page)
@@ -71,7 +76,7 @@ test('Formulas in app definition update the running app immediately', async ({ p
 })
 test('Invalid formula in app definition shows empty content in the running app until corrected', async ({ page })=> {
     await page.goto(runtimeRootUrl)
-    const appFrame = page.frame('appFrame') as Frame
+    const appFrame = await getAppFrame(page)
     expect(await appFrame.textContent('p >> nth=2')).toBe('Start your program here...')
     await openMainPage(page)
 
@@ -87,7 +92,7 @@ test('Invalid formula in app definition shows empty content in the running app u
 
 test('Global functions can be used in formulas', async ({ page })=> {
     await page.goto(runtimeRootUrl)
-    const appFrame = page.frame('appFrame') as Frame
+    const appFrame = await getAppFrame(page)
     expect(await appFrame.textContent('p >> nth=2')).toBe('Start your program here...')
 
     await openMainPage(page)
